feat: extract session state persistence into session.rs (#632)
Move SessionState, SessionLoad, load/save/GC, compartment helpers, and
related tests from main.rs into a dedicated session module. Reduces
main.rs from 3841 to 3142 lines (-699), continuing the #539 monolith
splitting effort. Ratchet ceiling lowered 3845 → 3145.

Closes #632
